+++ src/util/irep_hash.h
@@ -82,7 +82,8 @@ inline std::size_t basic_hash_combine<32>(
   std::size_t h1,
   std::size_t h2)
 {
-  return ROTL32(h1, 7)^h2;
+  return static_cast<std::size_t>(
+    ROTL32(static_cast<uint32_t>(h1), 7) ^ static_cast<uint32_t>(h2));
 }
 
 template<>
